refactor: replace `<Home />` by `<ConversorPage />` component
